refactor: absolute-style.scraper.ts 수정
