+++ src/notices/scrapers/relative-style/relative-style.scraper.ts
@@ -8,52 +8,21 @@
  * Latest Updated Date: 2025-05-18
  */
 
-import { Logger } from "@nestjs/common";
 import axios, { AxiosResponse } from 'axios';
 import { HTTP_STATUS } from "src/constants/http/http-status.constant";
 import { NotificationPayload } from "src/interfaces/notification-payload.interface";
-import { BaseScraper } from "../base.scraper";
+import { BaseScraper } from 'src/notices/scrapers/base.scraper';
 
 export abstract class RelativeStyleScraper extends BaseScraper {
-    protected logger: Logger;
-    protected configName: string;
-    protected noticeTypes: string[];
-    protected noticeTypeUrls: Record<string, string>;
-    protected noticeTypeQueryUrls: Record<string, string>;
-
     abstract fetchGeneralNotices(response: AxiosResponse<any>, queryUrl: string, noticeType: string): NotificationPayload[];
 
-    protected loadUrls(
-        noticeConfig: Record<string, { baseUrl: string; queryUrl: string }>,
-        key: 'baseUrl' | 'queryUrl'
-    ): Record<string, string> {
-        const results: Record<string, string> = {};
-
-        for (let i = 0; i < this.noticeTypes.length; i++) {
-            const noticeType = this.noticeTypes[i];
-            results[noticeType] = noticeConfig[noticeType]?.[key] || '';
-        }
-
-        return results;
-    }
-
-    async fetchAllNotices(): Promise<Record<string, NotificationPayload[]>> {
-        const results: Record<string, NotificationPayload[]> = {};
-
-        for (const noticeType of this.noticeTypes) {
-            try {
-                const notices: { general: NotificationPayload[] } = await this.fetchNotices(noticeType);
-                results[noticeType] = notices.general;
-            } catch (error) {
-                this.logger.error(`❌ ${noticeType} 공지사항 크롤링 실패:`, error.message);
-            }
-        }
-
-        return results;
-    }
-
-
-    async fetchNotices(noticeType: string): Promise<{ general: NotificationPayload[] }> {
+    /**
+     * 공지타입과 페이지 기반의 일반 공지사항을 크롤링하는 함수
+     * @param {string} noticeType - 공지타입
+     * @param {number} page - 페이지 번호
+     * @returns {Promise<{general: NotificationPayload[]}>} - 공지사항 객체 배열로 반환
+     */
+    async fetchNotices(noticeType: string, page: number): Promise<{ general: NotificationPayload[] }> {
         const baseUrl = this.noticeTypeUrls[noticeType];
         const queryUrl = this.noticeTypeQueryUrls[noticeType];
 
@@ -67,7 +36,7 @@ export abstract class RelativeStyleScraper extends BaseScraper {
             nameOption: '',
             onlyWriter: 'false',
             max: '10',
-            offset: '0',
+            offset: (page - 1).toString(),
         };
 
         try {
@@ -83,9 +52,5 @@ export abstract class RelativeStyleScraper extends BaseScraper {
         } catch (error) {
             throw new Error(`❌ ${noticeType}의 공지사항 크롤링 실패: ${error.message}`);
         }
-    }
-
-    getAllNoticeTypes(): string[] {
-        return this.noticeTypes;
     }
 }
